requirements.txt
Flask==3.1.3
requests==2.33.1
google-play-scraper==1.2.7
google-cloud-storage==3.10.1
PyJWT==2.12.1
python-dotenv==1.2.2
